refactor(sentry): consolidate all listeners into single EventListener
- Merge 11 individual listener files into one unified EventListener.php
- Maintain all existing functionality for exception handling and breadcrumb tracking
- Improve code organization and reduce file complexity
- Preserve configuration setup and event handling logic
- Support all framework events: AMQP, AsyncQueue, Command, Crontab, Database, Kafka, Redis, Request
