--- src/sentry/src/Listener/CommandExceptionListener.php
@@ -1,42 +1,0 @@
-<?php
-
-declare(strict_types=1);
-/**
- * This file is part of friendsofhyperf/components.
- *
- * @link     https://github.com/friendsofhyperf/components
- * @document https://github.com/friendsofhyperf/components/blob/main/README.md
- * @contact  huangdijia@gmail.com
- */
-
-namespace FriendsOfHyperf\Sentry\Listener;
-
-use Hyperf\Command\Event;
-
-class CommandExceptionListener extends CaptureExceptionListener
-{
-    public function listen(): array
-    {
-        return [
-            Event\BeforeHandle::class,
-            Event\FailToHandle::class,
-            Event\AfterExecute::class,
-        ];
-    }
-
-    /**
-     * @param Event\FailToHandle|Event\BeforeHandle|Event\AfterExecute|object $event
-     */
-    public function process(object $event): void
-    {
-        if (! $this->switcher->isEnable('command')) {
-            return;
-        }
-
-        match ($event::class) {
-            Event\FailToHandle::class => $this->captureException($event->getThrowable()),
-            Event\AfterExecute::class => $this->flushEvents(),
-            default => $this->setupSentrySdk(),
-        };
-    }
-}
