--- src/sentry/src/Listener/RedisCommandExecutedListener.php
@@ -1,57 +1,0 @@
-<?php
-
-declare(strict_types=1);
-/**
- * This file is part of friendsofhyperf/components.
- *
- * @link     https://github.com/friendsofhyperf/components
- * @document https://github.com/friendsofhyperf/components/blob/main/README.md
- * @contact  huangdijia@gmail.com
- */
-
-namespace FriendsOfHyperf\Sentry\Listener;
-
-use FriendsOfHyperf\Sentry\Integration;
-use FriendsOfHyperf\Sentry\Switcher;
-use Hyperf\Event\Contract\ListenerInterface;
-use Hyperf\Redis\Event\CommandExecuted;
-use Sentry\Breadcrumb;
-
-class RedisCommandExecutedListener implements ListenerInterface
-{
-    public function __construct(private Switcher $switcher)
-    {
-    }
-
-    public function listen(): array
-    {
-        return [
-            CommandExecuted::class,
-        ];
-    }
-
-    /**
-     * @param object|CommandExecuted $event
-     */
-    public function process(object $event): void
-    {
-        if (
-            ! $this->switcher->isBreadcrumbEnable('redis')
-            || ! $event instanceof CommandExecuted
-        ) {
-            return;
-        }
-
-        Integration::addBreadcrumb(new Breadcrumb(
-            Breadcrumb::LEVEL_INFO,
-            Breadcrumb::TYPE_DEFAULT,
-            'redis',
-            $event->command,
-            [
-                'arguments' => $event->parameters,
-                'result' => $event->result,
-                'duration' => $event->time * 1000,
-            ]
-        ));
-    }
-}
