--- src/sentry/src/Listener/SetRedisEventEnableListener.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-/**
- * This file is part of friendsofhyperf/components.
- *
- * @link     https://github.com/friendsofhyperf/components
- * @document https://github.com/friendsofhyperf/components/blob/main/README.md
- * @contact  huangdijia@gmail.com
- */
-
-namespace FriendsOfHyperf\Sentry\Listener;
-
-use Hyperf\Contract\ConfigInterface;
-use Hyperf\Event\Contract\ListenerInterface;
-use Hyperf\Framework\Event\BootApplication;
-
-class SetRedisEventEnableListener implements ListenerInterface
-{
-    public function __construct(protected ConfigInterface $config)
-    {
-    }
-
-    public function listen(): array
-    {
-        return [
-            BootApplication::class,
-        ];
-    }
-
-    public function process(object $event): void
-    {
-        if (! $this->config->has('redis')) {
-            return;
-        }
-
-        foreach ($this->config->get('redis', []) as $pool => $_) {
-            $this->config->set("redis.{$pool}.event.enable", true);
-        }
-    }
-}
